requirements.txt
# =====================================================================
# Dependencias de Biblioteca Virtual Django (versiones probadas)
# Instalar:  pip install -r requirements.txt
# Requiere:  Python 3.10+ (probado con 3.11)
# =====================================================================

# --- Framework web ---
Django==4.2.30
djangorestframework==3.17.1
django-cors-headers==4.9.0

# --- Servidor ASGI (para supervisor/uvicorn; opcional si usas runserver) ---
uvicorn[standard]==0.35.0

# --- MongoDB (el dominio vive en Atlas, no en SQLite) ---
pymongo==4.16.0
dnspython==2.8.0       # necesario para mongodb+srv:// (SRV records)

# --- Auth ---
PyJWT==2.12.1
bcrypt==4.1.3

# --- Utilidades ---
python-dotenv==1.2.2
requests==2.33.1

# --- Transitivas (se instalan solas, las pineamos por reproducibilidad) ---
asgiref==3.11.1
sqlparse==0.5.5
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
urllib3==2.6.3
